restructuring; build process; 1.7.0.pre5
