+++ Gemfile
@@ -1,11 +1,14 @@
+# frozen_string_literal: true
+
 source 'https://rubygems.org'
 
 # Specify your gem's dependencies in smarter_csv.gemspec
 gemspec
 
 gem "rake" # , "< 11"
+gem "rake-compiler"
+
 gem 'pry'
 
-group :test do
-  gem "rspec"
-end
+gem "rspec"
+gem "rubocop"
