+++ Rakefile
@@ -1,55 +1,21 @@
-#!/usr/bin/env rake
+# frozen_string_literal: true
+
 require "bundler/gem_tasks"
-require 'rubygems'
-require 'rake'
 require 'rspec/core/rake_task'
-require 'fileutils'
-require "rubygems/package_task"
 
-# require 'rake/extensiontask'
+RSpec::Core::RakeTask.new(:spec)
 
-# Rake::ExtensionTask.new('smarter_csv') do |ext|
-#   ext.ext_dir = 'ext/smarter_csv'
-# end
+require "rubocop/rake_task"
 
-# --------------------------------------------------------------------------
-# SMARTER_CSV_SPEC = Bundler.load_gemspec("smarter_csv.gemspec")
-# exttask = Rake::ExtensionTask.new('smarter_csv', SMARTER_CSV_SPEC) do |ext|
-#   ext.cross_compile = true
-#   ext.cross_platform = %w[x86-mingw32 x64-mingw32 x86-linux x86_64-linux]
-# end
+RuboCop::RakeTask.new
 
-require 'smarter_csv'
+require "rake/extensiontask"
 
-desc "Run RSpec"
-RSpec::Core::RakeTask.new do |t|
-  t.verbose = true
+task build: :compile
+
+Rake::ExtensionTask.new("smarter_csv") do |ext|
+  ext.ext_dir = "ext/smarter_csv"
 end
 
-task :test => :spec
-
-task :clean do
-  cd "ext/smarter_csv"
-  sh "rm -f Makefile"
-  sh "rm -f parse_csv_line.o"
-  sh "rm -f parse_csv_line.bundle"
-  cd "../.."
-end
-
-task :create_makefile do
-  cd "ext/smarter_csv"
-  sh "ruby extconf.rb"
-  cd "../.."
-  puts "\nPWD 0A: #{`pwd`}\n--------------------------------------\ndone"
-end
-
-task :compile do
-  puts "\n--------------------------------------\nCOMPILING...\n"
-  cd "ext/smarter_csv"
-  sh "ruby extconf.rb"
-  sh "make"
-  cd "../.."
-  puts "\nPWD 0B: #{`pwd`}\n--------------------------------------\ndone"
-end
-
-task default: [:compile, :test]
+# task default: %i[clobber compile spec rubocop]
+task default: %i[clobber compile spec]
