+++ ext/smarter_csv/extconf.rb
@@ -11,4 +11,4 @@ end
 
 CONFIG["optflags"] = "-O3"
 
-create_makefile('smarter_csv/parse_csv_line')
+create_makefile('smarter_csv/smarter_csv')
