+++ lib/smarter_csv/version.rb
@@ -1,3 +1,3 @@
 module SmarterCSV
-  VERSION = "1.7.0.pre4"
+  VERSION = "1.7.0.pre5"
 end
