+++ smarter_csv.gemspec
@@ -12,21 +12,32 @@ Gem::Specification.new do |spec|
   spec.homepage      = "https://github.com/tilo/smarter_csv"
   spec.license       = 'MIT'
 
-  spec.files         = Dir.glob('lib/**/*.rb') + Dir.glob('ext/**/*.{c, h, rb}')
-  spec.require_paths = ["lib", "ext"]
+  spec.metadata["homepage_uri"] = spec.homepage
+  spec.metadata["source_code_uri"] = spec.homepage
+  spec.metadata["changelog_uri"] = "https://github.com/tilo/smarter_csv/CHANGELOG.md"
+
+  # Specify which files should be added to the gem when it is released.
+  # The `git ls-files -z` loads the files in the RubyGem that have been added into git.
+  spec.files = Dir.chdir(__dir__) do
+    `git ls-files -z`.split("\x0").reject do |f|
+      (f == __FILE__) ||
+        f.match(%r{\A(?:(?:bin|test|spec|features)/|\.(?:git|travis|circleci)|appveyor)}) || f.match(/\.h\z/)
+    end
+  end
+  spec.test_files = spec.files.grep(%r{^(test|spec|features)/})
 
   spec.executables   = spec.files.grep(%r{^bin/}).map{ |f| File.basename(f) }
-  spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
+  spec.require_paths = ["lib"] # add ext here?
+  spec.extensions = ["ext/smarter_csv/extconf.rb"]
 
+
+  spec.add_development_dependency "awesome_print"
+  spec.add_development_dependency "codecov"
   spec.add_development_dependency "pry"
   spec.add_development_dependency "rspec"
+  spec.add_development_dependency "rubocop"
   spec.add_development_dependency "simplecov"
-  spec.add_development_dependency "codecov"
-  spec.add_development_dependency "awesome_print"
 
-  spec.metadata["homepage_uri"] = spec.homepage
-
-  spec.extensions << "ext/smarter_csv/extconf.rb"
 
   puts "FILES: #{spec.files.inspect}"
 end
